# openapi.yaml
openapi: 3.0.3
info:
  title: Saukele API
  description: Digital Wedding Registry & Kazakh Tradition Logistics Platform
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
...
  /weddings:
    post:
      tags: [Weddings]
      summary: Create a wedding profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [brideName, groomName, weddingDate, location, weddingType]
              properties:
                brideName:
                  type: string
                  example: Aisha Bekova
                groomName:
                  type: string
                  example: Daniyar Seitkali
                weddingDate:
                  type: string
                  format: date
                  example: 2025-06-15
                location:
                  type: string
                  example: Almaty
                weddingType:
                  type: string
                  enum: [CITY, VILLAGE]
                  example: CITY
      responses:
        '201':
          description: Wedding profile created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: uuid-here
                  inviteCode:
                    type: string
                    example: uuid-invite-code
                  brideName:
                    type: string
                    example: Aisha Bekova
                  groomName:
                    type: string
                    example: Daniyar Seitkali
                  weddingDate:
                    type: string
                    example: 2025-06-15
                  weddingType:
                    type: string
                    example: CITY
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '409':
          description: Wedding profile already exists
        '422':
          description: Validation error
        '500':
          description: Internal server error
...
  /gifts:
    post:
      tags: [Gifts]
      summary: Add a gift to the registry
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [weddingProfileId, name, price]
              properties:
                weddingProfileId:
                  type: string
                  example: uuid-here
                vendorId:
                  type: string
                  example: uuid-here
                name:
                  type: string
                  example: Handmade Kilem
                description:
                  type: string
                  example: Traditional Kazakh carpet
                price:
                  type: number
                  example: 150000
                isPoolGift:
                  type: boolean
                  example: true
                targetAmount:
                  type: number
                  example: 150000
                privacy:
                  type: string
                  enum: [PUBLIC, FAMILY_ONLY]
                  example: PUBLIC
                isFragile:
                  type: boolean
                  example: true
                deadlineAt:
                  type: string
                  format: date-time
                  example: 2025-06-01T00:00:00Z
      responses:
        '201':
          description: Gift added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: uuid-here
                  name:
                    type: string
                    example: Handmade Kilem
                  price:
                    type: number
                    example: 150000
                  status:
                    type: string
                    example: AVAILABLE
                  isPoolGift:
                    type: boolean
                    example: true
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '422':
          description: Validation error
        '500':
          description: Internal server error
    get:
      tags: [Gifts]
      summary: List all gifts for a wedding registry
      security:
        - BearerAuth: []
      parameters:
        - name: weddingProfileId
          in: query
          required: true
          schema:
            type: string
            example: uuid-here
        - name: status
          in: query
          schema:
            type: string
            enum: [AVAILABLE, PARTIALLY_FUNDED, FUNDED, PURCHASED, DELIVERED]
            example: AVAILABLE
        - name: cursor
          in: query
          schema:
            type: string
          description: Cursor for pagination
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            example: 20
      responses:
        '200':
          description: List of gifts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: uuid-here
                        name:
                          type: string
                          example: Handmade Kilem
                        price:
                          type: number
                          example: 150000
                        status:
                          type: string
                          example: AVAILABLE
                  nextCursor:
                    type: string
                    example: uuid-next-cursor
        '401':
          description: Unauthorized
        '404':
          description: Wedding profile not found
        '500':
          description: Internal server error
...
  /contributions:
    post:
      tags: [Contributions]
      summary: Make a contribution to a pool gift
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [giftItemId, amountOriginal, currency]
              properties:
                giftItemId:
                  type: string
                  example: uuid-here
                amountOriginal:
                  type: number
                  example: 5000
                currency:
                  type: string
                  enum: [KZT, USD]
                  example: KZT
                idempotencyKey:
                  type: string
                  example: unique-key-123
      responses:
        '201':
          description: Contribution created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: uuid-here
                  amountOriginal:
                    type: number
                    example: 5000
                  currency:
                    type: string
                    example: KZT
                  amountKzt:
                    type: number
                    example: 5000
                  exchangeRate:
                    type: number
                    example: 1
                  lockedAt:
                    type: string
                    example: 2025-01-01T00:00:00Z
                  status:
                    type: string
                    example: PENDING
        '400':
          description: Contribution exceeds remaining pool amount
        '401':
          description: Unauthorized
        '404':
          description: Gift not found
        '409':
          description: Duplicate idempotency key
        '422':
          description: Validation error
        '500':
          description: Internal server error
    get:
      tags: [Contributions]
      summary: List contributions for a gift
      security:
        - BearerAuth: []
      parameters:
        - name: giftItemId
          in: query
          required: true
          schema:
            type: string
            example: uuid-here
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, CONFIRMED, REFUNDED]
            example: CONFIRMED
        - name: cursor
          in: query
          schema:
            type: string
          description: Cursor for pagination
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            example: 20
      responses:
        '200':
          description: List of contributions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: uuid-here
                        amountOriginal:
                          type: number
                          example: 5000
                        currency:
                          type: string
                          example: KZT
                        status:
                          type: string
                          example: CONFIRMED
                  nextCursor:
                    type: string
                    example: uuid-next-cursor
        '401':
          description: Unauthorized
        '404':
          description: Gift not found
        '500':
          description: Internal server error
...
  /vendors:
    post:
      tags: [Vendors]
      summary: Register as a vendor
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shopName, phone]
              properties:
                shopName:
                  type: string
                  example: Aisha Handmade
                description:
                  type: string
                  example: Traditional Kazakh handmade items
                phone:
                  type: string
                  example: +77001234567
      responses:
        '201':
          description: Vendor registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: uuid-here
                  shopName:
                    type: string
                    example: Aisha Handmade
                  status:
                    type: string
                    example: PENDING
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '409':
          description: Vendor profile already exists
        '422':
          description: Validation error
        '500':
          description: Internal server error
    get:
      tags: [Vendors]
      summary: List all approved vendors
      security:
        - BearerAuth: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
          description: Cursor for pagination
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            example: 20
      responses:
        '200':
          description: List of vendors
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: uuid-here
                        shopName:
                          type: string
                          example: Aisha Handmade
                        status:
                          type: string
                          example: APPROVED
                  nextCursor:
                    type: string
                    example: uuid-next-cursor
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
...
  /orders:
    post:
      tags: [Orders]
      summary: Create an order for a funded gift
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [giftItemId, vendorId, totalAmount]
              properties:
                giftItemId:
                  type: string
                  example: uuid-here
                vendorId:
                  type: string
                  example: uuid-here
                totalAmount:
                  type: number
                  example: 150000
                notes:
                  type: string
                  example: Please embroider initials A&D
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: uuid-here
                  status:
                    type: string
                    example: PLACED
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '404':
          description: Gift not found
        '409':
          description: Order already exists for this gift
        '422':
          description: Validation error
        '500':
          description: Internal server error
    get:
      tags: [Orders]
      summary: List all orders
      security:
        - BearerAuth: []
      parameters:
        - name: vendorId
          in: query
          schema:
            type: string
            example: uuid-here
        - name: status
          in: query
          schema:
            type: string
            enum: [PLACED, CONFIRMED, READY_FOR_PICKUP, PICKED_UP, DELIVERED]
            example: PLACED
        - name: cursor
          in: query
          schema:
            type: string
          description: Cursor for pagination
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            example: 20
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: uuid-here
                        giftItemId:
                          type: string
                          example: uuid-here
                        vendorId:
                          type: string
                          example: uuid-here
                        totalAmount:
                          type: number
                          example: 150000
                        status:
                          type: string
                          example: PLACED
                  nextCursor:
                    type: string
                    example: uuid-next-cursor
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
...
  /family-members:
    post:
      tags: [Weddings]
      summary: Add a family member to wedding guest list
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [weddingProfileId, guestId, kinshipType, giftTier]
              properties:
                weddingProfileId:
                  type: string
                  example: uuid-here
                guestId:
                  type: string
                  example: uuid-here
                kinshipType:
                  type: string
                  enum: [ATA_ANA, AGA_INI, ZHIEN, QUDA, OTHER]
                  example: ATA_ANA
                giftTier:
                  type: string
                  enum: [TIER_1, TIER_2, TIER_3, TIER_4, TIER_5]
                  example: TIER_1
                parentId:
                  type: string
                  example: uuid-here
      responses:
        '201':
          description: Family member added
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: uuid-here
                  kinshipType:
                    type: string
                    example: ATA_ANA
                  giftTier:
                    type: string
                    example: TIER_1
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '409':
          description: Guest already in family list
        '422':
          description: Validation error
        '500':
          description: Internal server error
    get:
      tags: [Weddings]
      summary: List all family members for a wedding
      security:
        - BearerAuth: []
      parameters:
        - name: weddingProfileId
          in: query
          required: true
          schema:
            type: string
            example: uuid-here
        - name: cursor
          in: query
          schema:
            type: string
          description: Cursor for pagination
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            example: 20
      responses:
        '200':
          description: List of family members
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: uuid-here
                        guestId:
                          type: string
                          example: uuid-here
                        kinshipType:
                          type: string
                          example: ATA_ANA
                        giftTier:
                          type: string
                          example: TIER_1
                  nextCursor:
                    type: string
                    example: uuid-next-cursor
        '401':
          description: Unauthorized
        '404':
          description: Wedding profile not found
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT